+++ sdk/resourcemanager/datafactory/armdatafactory/models_serde.go
@@ -11696,6 +11696,7 @@ func (c CommonDataServiceForAppsLinkedServiceTypeProperties) MarshalJSON() ([]by
 	objectMap := make(map[string]any)
 	populateAny(objectMap, "authenticationType", c.AuthenticationType)
 	populateAny(objectMap, "deploymentType", c.DeploymentType)
+	populateAny(objectMap, "domain", c.Domain)
 	populate(objectMap, "encryptedCredential", c.EncryptedCredential)
 	populateAny(objectMap, "hostName", c.HostName)
 	populateAny(objectMap, "organizationName", c.OrganizationName)
@@ -11723,6 +11724,9 @@ func (c *CommonDataServiceForAppsLinkedServiceTypeProperties) UnmarshalJSON(data
 			delete(rawMsg, key)
 		case "deploymentType":
 			err = unpopulate(val, "DeploymentType", &c.DeploymentType)
+			delete(rawMsg, key)
+		case "domain":
+			err = unpopulate(val, "Domain", &c.Domain)
 			delete(rawMsg, key)
 		case "encryptedCredential":
 			err = unpopulate(val, "EncryptedCredential", &c.EncryptedCredential)
@@ -12266,6 +12270,41 @@ func (c *ConnectionStateProperties) UnmarshalJSON(data []byte) error {
 			delete(rawMsg, key)
 		case "status":
 			err = unpopulate(val, "Status", &c.Status)
+			delete(rawMsg, key)
+		}
+		if err != nil {
+			return fmt.Errorf("unmarshalling type %T: %v", c, err)
+		}
+	}
+	return nil
+}
+
+// MarshalJSON implements the json.Marshaller interface for type ContinuationSettingsReference.
+func (c ContinuationSettingsReference) MarshalJSON() ([]byte, error) {
+	objectMap := make(map[string]any)
+	populateAny(objectMap, "continuationTtlInMinutes", c.ContinuationTTLInMinutes)
+	populateAny(objectMap, "customizedCheckpointKey", c.CustomizedCheckpointKey)
+	populateAny(objectMap, "idleCondition", c.IdleCondition)
+	return json.Marshal(objectMap)
+}
+
+// UnmarshalJSON implements the json.Unmarshaller interface for type ContinuationSettingsReference.
+func (c *ContinuationSettingsReference) UnmarshalJSON(data []byte) error {
+	var rawMsg map[string]json.RawMessage
+	if err := json.Unmarshal(data, &rawMsg); err != nil {
+		return fmt.Errorf("unmarshalling type %T: %v", c, err)
+	}
+	for key, val := range rawMsg {
+		var err error
+		switch key {
+		case "continuationTtlInMinutes":
+			err = unpopulate(val, "ContinuationTTLInMinutes", &c.ContinuationTTLInMinutes)
+			delete(rawMsg, key)
+		case "customizedCheckpointKey":
+			err = unpopulate(val, "CustomizedCheckpointKey", &c.CustomizedCheckpointKey)
+			delete(rawMsg, key)
+		case "idleCondition":
+			err = unpopulate(val, "IdleCondition", &c.IdleCondition)
 			delete(rawMsg, key)
 		}
 		if err != nil {
@@ -18020,6 +18059,7 @@ func (d DynamicsCrmLinkedServiceTypeProperties) MarshalJSON() ([]byte, error) {
 	populateAny(objectMap, "authenticationType", d.AuthenticationType)
 	populate(objectMap, "credential", d.Credential)
 	populateAny(objectMap, "deploymentType", d.DeploymentType)
+	populateAny(objectMap, "domain", d.Domain)
 	populate(objectMap, "encryptedCredential", d.EncryptedCredential)
 	populateAny(objectMap, "hostName", d.HostName)
 	populateAny(objectMap, "organizationName", d.OrganizationName)
@@ -18050,6 +18090,9 @@ func (d *DynamicsCrmLinkedServiceTypeProperties) UnmarshalJSON(data []byte) erro
 			delete(rawMsg, key)
 		case "deploymentType":
 			err = unpopulate(val, "DeploymentType", &d.DeploymentType)
+			delete(rawMsg, key)
+		case "domain":
+			err = unpopulate(val, "Domain", &d.Domain)
 			delete(rawMsg, key)
 		case "encryptedCredential":
 			err = unpopulate(val, "EncryptedCredential", &d.EncryptedCredential)
@@ -18402,6 +18445,7 @@ func (d DynamicsLinkedServiceTypeProperties) MarshalJSON() ([]byte, error) {
 	populateAny(objectMap, "authenticationType", d.AuthenticationType)
 	populate(objectMap, "credential", d.Credential)
 	populateAny(objectMap, "deploymentType", d.DeploymentType)
+	populateAny(objectMap, "domain", d.Domain)
 	populate(objectMap, "encryptedCredential", d.EncryptedCredential)
 	populateAny(objectMap, "hostName", d.HostName)
 	populateAny(objectMap, "organizationName", d.OrganizationName)
@@ -18432,6 +18476,9 @@ func (d *DynamicsLinkedServiceTypeProperties) UnmarshalJSON(data []byte) error {
 			delete(rawMsg, key)
 		case "deploymentType":
 			err = unpopulate(val, "DeploymentType", &d.DeploymentType)
+			delete(rawMsg, key)
+		case "domain":
+			err = unpopulate(val, "Domain", &d.Domain)
 			delete(rawMsg, key)
 		case "encryptedCredential":
 			err = unpopulate(val, "EncryptedCredential", &d.EncryptedCredential)
@@ -19277,6 +19324,7 @@ func (e *ExecuteDataFlowActivity) UnmarshalJSON(data []byte) error {
 func (e ExecuteDataFlowActivityTypeProperties) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]any)
 	populate(objectMap, "compute", e.Compute)
+	populate(objectMap, "continuationSettings", e.ContinuationSettings)
 	populateAny(objectMap, "continueOnError", e.ContinueOnError)
 	populate(objectMap, "dataFlow", e.DataFlow)
 	populate(objectMap, "integrationRuntime", e.IntegrationRuntime)
@@ -19298,6 +19346,9 @@ func (e *ExecuteDataFlowActivityTypeProperties) UnmarshalJSON(data []byte) error
 		switch key {
 		case "compute":
 			err = unpopulate(val, "Compute", &e.Compute)
+			delete(rawMsg, key)
+		case "continuationSettings":
+			err = unpopulate(val, "ContinuationSettings", &e.ContinuationSettings)
 			delete(rawMsg, key)
 		case "continueOnError":
 			err = unpopulate(val, "ContinueOnError", &e.ContinueOnError)
@@ -19514,6 +19565,7 @@ func (e *ExecutePipelineActivityTypeProperties) UnmarshalJSON(data []byte) error
 func (e ExecutePowerQueryActivityTypeProperties) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]any)
 	populate(objectMap, "compute", e.Compute)
+	populate(objectMap, "continuationSettings", e.ContinuationSettings)
 	populateAny(objectMap, "continueOnError", e.ContinueOnError)
 	populate(objectMap, "dataFlow", e.DataFlow)
 	populate(objectMap, "integrationRuntime", e.IntegrationRuntime)
@@ -19537,6 +19589,9 @@ func (e *ExecutePowerQueryActivityTypeProperties) UnmarshalJSON(data []byte) err
 		switch key {
 		case "compute":
 			err = unpopulate(val, "Compute", &e.Compute)
+			delete(rawMsg, key)
+		case "continuationSettings":
+			err = unpopulate(val, "ContinuationSettings", &e.ContinuationSettings)
 			delete(rawMsg, key)
 		case "continueOnError":
 			err = unpopulate(val, "ContinueOnError", &e.ContinueOnError)
@@ -42034,6 +42089,7 @@ func (s SQLServerLinkedServiceTypeProperties) MarshalJSON() ([]byte, error) {
 	populateAny(objectMap, "connectRetryInterval", s.ConnectRetryInterval)
 	populateAny(objectMap, "connectTimeout", s.ConnectTimeout)
 	populateAny(objectMap, "connectionString", s.ConnectionString)
+	populate(objectMap, "credential", s.Credential)
 	populateAny(objectMap, "database", s.Database)
 	populateAny(objectMap, "encrypt", s.Encrypt)
 	populate(objectMap, "encryptedCredential", s.EncryptedCredential)
@@ -42086,6 +42142,9 @@ func (s *SQLServerLinkedServiceTypeProperties) UnmarshalJSON(data []byte) error 
 			delete(rawMsg, key)
 		case "connectionString":
 			err = unpopulate(val, "ConnectionString", &s.ConnectionString)
+			delete(rawMsg, key)
+		case "credential":
+			err = unpopulate(val, "Credential", &s.Credential)
 			delete(rawMsg, key)
 		case "database":
 			err = unpopulate(val, "Database", &s.Database)
@@ -49774,6 +49833,7 @@ func (s SnowflakeExportCopyCommand) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]any)
 	populate(objectMap, "additionalCopyOptions", s.AdditionalCopyOptions)
 	populate(objectMap, "additionalFormatOptions", s.AdditionalFormatOptions)
+	populateAny(objectMap, "storageIntegration", s.StorageIntegration)
 	objectMap["type"] = "SnowflakeExportCopyCommand"
 	if s.AdditionalProperties != nil {
 		for key, val := range s.AdditionalProperties {
@@ -49797,6 +49857,9 @@ func (s *SnowflakeExportCopyCommand) UnmarshalJSON(data []byte) error {
 			delete(rawMsg, key)
 		case "additionalFormatOptions":
 			err = unpopulate(val, "AdditionalFormatOptions", &s.AdditionalFormatOptions)
+			delete(rawMsg, key)
+		case "storageIntegration":
+			err = unpopulate(val, "StorageIntegration", &s.StorageIntegration)
 			delete(rawMsg, key)
 		case "type":
 			err = unpopulate(val, "Type", &s.Type)
@@ -49824,6 +49887,7 @@ func (s SnowflakeImportCopyCommand) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]any)
 	populate(objectMap, "additionalCopyOptions", s.AdditionalCopyOptions)
 	populate(objectMap, "additionalFormatOptions", s.AdditionalFormatOptions)
+	populateAny(objectMap, "storageIntegration", s.StorageIntegration)
 	objectMap["type"] = "SnowflakeImportCopyCommand"
 	if s.AdditionalProperties != nil {
 		for key, val := range s.AdditionalProperties {
@@ -49847,6 +49911,9 @@ func (s *SnowflakeImportCopyCommand) UnmarshalJSON(data []byte) error {
 			delete(rawMsg, key)
 		case "additionalFormatOptions":
 			err = unpopulate(val, "AdditionalFormatOptions", &s.AdditionalFormatOptions)
+			delete(rawMsg, key)
+		case "storageIntegration":
+			err = unpopulate(val, "StorageIntegration", &s.StorageIntegration)
 			delete(rawMsg, key)
 		case "type":
 			err = unpopulate(val, "Type", &s.Type)